templat fini a 80%

+++ XX_Project_location_voitures/src/main/java/com/adaming/tpLocVoitures/dao/IClientDAO.java
@@ -11,5 +11,5 @@ import com.adaming.tpLocVoitures.entities.Reservation;
  */
 public interface IClientDAO extends IGenericDAO<Client> {
 
-	List<Reservation> getAllReservations(Integer idClient);
+	List<Reservation> getAllReservations(Long idClient);
 }
