+++ XX_Project_location_voitures/src/main/java/com/adaming/tpLocVoitures/dao/IVoitureDAO.java
@@ -11,5 +11,5 @@ import com.adaming.tpLocVoitures.entities.Voiture;
  */
 public interface IVoitureDAO extends IGenericDAO<Voiture> {
 
-	List<Reservation> getAllReservations(Integer idVoiture);
+	List<Reservation> getAllReservations(Long idVoiture);
 }
